feat(services): Add analytics and workout management services
- Implement AnalyticsCalculationService for workout metrics
- Add SomeDuplicateEntryException for duplicate entry handling
- Enhance WorkoutMapperService with data transformation utilities
- Add WorkoutPlanController for workout plan REST endpoints
- Implement comprehensive workout and analytics processing

+++ src/main/java/com/muscledia/workout_service/service/WorkoutMapperService.java
@@ -131,7 +131,7 @@ public class WorkoutMapperService {
         return exercises.stream()
                 .filter(exercise -> exercise.getWeight() != null && exercise.getReps() != null
                         && exercise.getSets() != null)
-                .map(exercise -> exercise.getWeight()
+                .map(exercise -> BigDecimal.valueOf(exercise.getWeight()) // FIX: Convert Double to BigDecimal
                         .multiply(BigDecimal.valueOf(exercise.getReps()))
                         .multiply(BigDecimal.valueOf(exercise.getSets())))
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
